refactor: remove watch on session list
