+++ app/javascript/angular/code/controllers/sessions_list_ctrl.js
@@ -5,9 +5,7 @@ angular
   .controller("SessionsListCtrl", [
     "$scope",
     "params",
-    "sensors",
     "$window",
-    "drawSession",
     "sessionsUtils",
     "map",
     SessionsListCtrl
